[bugfix] Fixed bug where collections update _byId too late
When setting attributes and thus clearing a relation of a model, the related collections themselves cannot find the model if it got a new id within the set-function. This is now fixed by triggering change:idAttribute as soon as the id has changed, thus forcing all owning collections to update their _byId-lookup-shizzl.